+++ multi-app-workflows/java/ai-recommendation-service/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>io.dapr.quickstarts</groupId>
+    <groupId>io.dapr.samples</groupId>
     <artifactId>ai-recommendation-service</artifactId>
     <version>1.0.0</version>
     <packaging>jar</packaging>
@@ -52,7 +52,7 @@
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>io.dapr.quickstarts.workflows.AIRecommendationServiceApplication</mainClass>
+                                    <mainClass>io.dapr.samples.workflows.AIRecommendationServiceApplication</mainClass>
                                 </transformer>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                             </transformers>
