+++ multi-app-workflows/java/ai-recommendation-service/src/main/java/io/dapr/samples/workflows/AIRecommendationServiceApplication.java
@@ -1,8 +1,8 @@
-package io.dapr.quickstarts.workflows;
+package io.dapr.samples.workflows;
 
 import io.dapr.workflows.runtime.WorkflowRuntime;
 import io.dapr.workflows.runtime.WorkflowRuntimeBuilder;
-import io.dapr.quickstarts.workflows.activities.ValidatePaymentMethodActivity;
+import io.dapr.samples.workflows.activities.GeneratePersonalizedRecommendationsActivity;
 import com.sun.net.httpserver.HttpServer;
 import com.sun.net.httpserver.HttpHandler;
 import com.sun.net.httpserver.HttpExchange;
@@ -11,28 +11,28 @@ import java.io.OutputStream;
 import java.net.InetSocketAddress;
 
 /**
- * PaymentServiceWorker - registers only the ValidatePaymentMethodActivity.
- * This app will handle multi-app activity calls from the main workflow (written in Go).
+ * AIRecommendationServiceWorker - registers only the GeneratePersonalizedRecommendationsActivity.
+ * This activity is called multi-app from the main workflow (written in Go).
  */
-public class PaymentServiceApplication {
+public class AIRecommendationServiceApplication {
 
     public static void main(String[] args) throws Exception {
-        System.out.println("=== Starting PaymentServiceWorker (ValidatePaymentMethodActivity) ===");
+        System.out.println("=== Starting AIRecommendationServiceWorker (GeneratePersonalizedRecommendationsActivity) ===");
         
-        // Start HTTP server on port 50002
-        HttpServer server = HttpServer.create(new InetSocketAddress(50002), 0);
+        // Start HTTP server on port 50004
+        HttpServer server = HttpServer.create(new InetSocketAddress(50004), 0);
         server.createContext("/health", new HealthHandler());
         server.setExecutor(null);
         server.start();
-        System.out.println("HTTP server started on port 50002");
+        System.out.println("HTTP server started on port 50004");
 
         WorkflowRuntimeBuilder builder = new WorkflowRuntimeBuilder()
-                .registerActivity(ValidatePaymentMethodActivity.class);
+                .registerActivity(GeneratePersonalizedRecommendationsActivity.class);
 
         // Build and start the workflow runtime
         try (WorkflowRuntime runtime = builder.build()) {
-            System.out.println("PaymentServiceWorker started - registered ValidatePaymentMethodActivity only");
-            System.out.println("Payment Service is ready to receive multi-app activity calls...");
+            System.out.println("AIRecommendationServiceWorker started - registered GeneratePersonalizedRecommendationsActivity only");
+            System.out.println("AI Recommendation Service is ready to receive multi-app activity calls...");
             System.out.println("Waiting for multi-app activity calls...");
             runtime.start();
         }
@@ -41,7 +41,7 @@ public class PaymentServiceApplication {
     static class HealthHandler implements HttpHandler {
         @Override
         public void handle(HttpExchange exchange) throws IOException {
-            String response = "Payment Service is running";
+            String response = "AI Recommendation Service is running";
             exchange.sendResponseHeaders(200, response.length());
             OutputStream os = exchange.getResponseBody();
             os.write(response.getBytes());
