+++ multi-app-workflows/java/inventory-service/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>io.dapr.quickstarts</groupId>
+    <groupId>io.dapr.samples</groupId>
     <artifactId>inventory-service</artifactId>
     <version>1.0.0</version>
     <packaging>jar</packaging>
@@ -52,7 +52,7 @@
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>io.dapr.quickstarts.workflows.InventoryServiceApplication</mainClass>
+                                    <mainClass>io.dapr.samples.workflows.InventoryServiceApplication</mainClass>
                                 </transformer>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                             </transformers>
